Add a quick cson build for atom-tachyons
As a potential way to help the atom-tachyons keep
their package up to date I've added a cson build
script that generates most of the output that the
atom extension ecosystem likes.

This can be run with

    npm run cson

It will output everything to stdout. We may want
to perhaps write to a file instead?

This comes with the caveat that it doesn't include
the string templating syntax found here
https://github.com/calweb/atom-tachyons/blob/1f956950e5d1982fe0c2a2d157fa694e60bee78d/snippets/components.cson#L6

***

Related calweb/atom-tachyons#1